pyproject.toml
[tool.poetry]
name = "act-project-"
version = "0.1.0"
description = ""
authors = ["Kimmyeonghyeon0 <Kimmyeonghyeom0>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
black = "^24.3.0"
isort = "^5.13.2"
mypy = "^1.9.0"
django-stubs = "^5.1.1"
pillow = "^11.0.0"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.3.1"
social-auth-app-django = "^5.4.2"
djangorestframework = "^3.15.2"
django-allauth = "^65.3.0"
dj-rest-auth = "^7.0.0"
drf-spectacular = "^0.28.0"
python-decouple = "^3.8"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 88
known_third_party = ["django", "rest_framework"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

